#461 | Cleaning up EditCategoryNameController
